+++ CHANGELOG.md
@@ -959,3 +959,4 @@
 - [2026-05-06 19:17] test(navigation): implement widget integration tests for navigation
 - [2026-05-06 19:54] fix(security): prevent double tap duplicate action in security
 - [2026-05-06 20:24] feat(analytics): add async analytics stream provider
+- [2026-05-06 20:49] feat(cache): add bi-directional event bus handler for cache